Fix for add/subtract when crossing a tz offset
When using 'add' or 'subtract' (or maybe other) moment.js functions, result was wrong if date passes through tz offset change.  See discussion in issue #28 for more details.

This commit includes a small suite of tests, along with a small fix in the spirit of that suggested by @icambron [here][1]: "If we're changing a date or month, just [...] change the offset without touching the time."

[1]: https://github.com/moment/moment-timezone/issues/28#issuecomment-25409441

This requires an associated change to the moment project, to support this second argument (though it will not fail any harder than it currently does, if running with an older version of moment).