Add console error when moment.momentProperties undefined
I had an error where it said _z is undefined.  This took me about 5 minutes to breakpoint the code, which lead me back to the site where I found moment 2.6 was required.  Because this threw an uncaught error other items broke like not being able to load data.  By having if statement you get the benefit of

#1 No uncaught exception where user is confused
#2 User gets a helpful message in their console and a link to correct the dependency