Restructure zone getter to allow for lazy unpacking.
`getZone` first checks if a zone is already parsed and returns it if so.
`getZone` first checks if there is a packed string available to use. If there is, it unpacks it and returns the new zone.
`getZone` then checks if there is link defined for that name. If there is, it calls `getZone` with the linked name and uses that to populate the data.