Initial work on guessing timezone.
Add tests for guessing the current zone.

Add tests ensuring that each zone is represented in the guesses.

Test that London, New York, and Los Angeles all have guesses and are
not equal.

Attach 'guess' property to the packed zone string.

Ensure that guessed zones are never made into links.