Remove invalid jspm configuration
The NPM package is not published with `builds/moment-timezone-with-data` so it's an error to specify it as the main file. It also appears to confuse jspm which tries to import `moment-timezone.js` instead of `index.js` (which is required as it loads the necessary data). Removing this invalid configuration will make jspm import the correct file (`index.js`).