Bugfix: data: ignore zdump output lacking utc/local time
In some extreme cases (with dates very far in the past/future) zdump
fails to represent the timestamp as a string, printing

    -67768040609740801 (gmtime failed)

or

    -67768040609746397 (localtime failed)

in this case, where the first number is the linux epoch timestamp.

These still got parsed by the script resulting in null values in the
json file.

So skip those lines and avoid the null values in the collected json
file.