Bugfix: data: Fix dependency order
Both collect and dedupe use meta to inject countries and population in
resulting json, but the meta task was run afterwards.

When used with latest, it just used the previous meta latest file, which
is a bug, and when used with a custom version (that wasn't generated
before) the pipeline crashed right away.

Move meta generation right after download task, as it only depends on
raw files (iso3166.tab and zone*.tab).