data: automatically create data/*/VERSION.json for latest
Commited files under data/ folder contain metadata, packed and unpacked
data for all released versions. Releases are normally done with `grunt
release` which calls `grunt data` which uses version latest, so a manual
copy of `data/*/latest.json` to the actual latest version was required
(or forgotten).

So change the pipeline to automatically generate the correctly-versioned
files in addition to latest.

Also move version parsing to data-meta, where it belongs.