build: Use -V flag for zdump and log warning for old versions
The `-V` flag was introduced 10 years ago but macOS uses a very old version of
`zdump` that doesn't support it. The versions of `zic` and `zdump` shipped with
macOS only support 32-bit data files.

A while ago the `data-zdump` task was changed from `-V` to `-v` to support
building on macOS, but that also switched to a far more verbose output format
on Linux machines. The best solution is to use `-v` for macOS but `-V` for
newer versions of `zdump`.

This also logs some warnings when `zic` and `zdump` are outdated.
The best solution for building the data on macOS currently is to manually
compile `zic` and `zdump` from the tzcode source.